requirements.txt
fastapi==0.88.0
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0
sqlalchemy==1.4.48
alembic==1.10.4
# psycopg2-binary==2.9.6  # Uncomment if using PostgreSQL
